Adopt a different approach to doing calculations.
Refactor the code to use "tables," "operators" and "arrays." Tables
contain column data and are wrappers round Pandas dataframes.
Calculations are made by applying operators:

    fractions = operator.transmitted_fraction() * materials

The result of a combination of an operator and a table is a table.
Operators can also be combined and then later applied to a table:

    op = operator.escaping_photons() * operator.transmitted_fraction()
    photons = op * photon_counts * system

Arrays are used to keep track of multiple scenarios.  An array
can contain a list of operations or a list of tables.  It is used
in expressions interchangeably with tables and operations, and
calculations or merges are carried out over all of the objects
within them.